moving createDates code into its own file
relates to #1
